liana-ui: split palette in liana + liana_business
