--- liana-ui/src/theme/palette.rs
@@ -1,1022 +1,0 @@
-use crate::color;
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Palette {
-    pub general: General,
-    pub text: Text,
-    pub buttons: Buttons,
-    pub cards: Cards,
-    pub banners: Banners,
-    pub badges: Badges,
-    pub pills: Pills,
-    pub notifications: Notifications,
-    pub text_inputs: TextInputs,
-    pub checkboxes: Checkboxes,
-    pub radio_buttons: RadioButtons,
-    pub sliders: Sliders,
-    pub progress_bars: ProgressBars,
-    pub rule: iced::Color,
-    pub pane_grid: PaneGrid,
-    pub togglers: Togglers,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Text {
-    pub primary: iced::Color,
-    pub secondary: iced::Color,
-    pub warning: iced::Color,
-    pub success: iced::Color,
-    pub error: iced::Color,
-    pub accent: iced::Color,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct General {
-    pub background: iced::Color,
-    pub foreground: iced::Color,
-    pub scrollable: iced::Color,
-    pub accent: iced::Color,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Buttons {
-    pub transparent: Button,
-    pub transparent_border: Button,
-    pub primary: Button,
-    pub secondary: Button,
-    pub destructive: Button,
-    pub container: Button,
-    pub container_border: Button,
-    pub menu: Button,
-    pub tab: Button,
-    pub link: Button,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Button {
-    pub active: ButtonPalette,
-    pub hovered: ButtonPalette,
-    pub pressed: Option<ButtonPalette>,
-    pub disabled: Option<ButtonPalette>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct ButtonPalette {
-    pub background: iced::Color,
-    pub text: iced::Color,
-    pub border: Option<iced::Color>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Containers {
-    pub border: ContainerPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct ContainerPalette {
-    pub background: iced::Color,
-    pub text: Option<iced::Color>,
-    pub border: Option<iced::Color>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Cards {
-    pub simple: ContainerPalette,
-    pub modal: ContainerPalette,
-    pub border: ContainerPalette,
-    pub invalid: ContainerPalette,
-    pub warning: ContainerPalette,
-    pub error: ContainerPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Banners {
-    pub network: ContainerPalette,
-    pub warning: ContainerPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Badges {
-    pub simple: ContainerPalette,
-    pub bitcoin: ContainerPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Pills {
-    pub simple: ContainerPalette,
-    pub primary: ContainerPalette,
-    pub success: ContainerPalette,
-    pub warning: ContainerPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Notifications {
-    pub pending: ContainerPalette,
-    pub error: ContainerPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct TextInputs {
-    pub primary: TextInput,
-    pub invalid: TextInput,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct TextInput {
-    pub active: TextInputPalette,
-    pub disabled: TextInputPalette,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct TextInputPalette {
-    pub background: iced::Color,
-    pub icon: iced::Color,
-    pub placeholder: iced::Color,
-    pub value: iced::Color,
-    pub selection: iced::Color,
-    pub border: Option<iced::Color>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Checkboxes {
-    pub icon: iced::Color,
-    pub text: iced::Color,
-    pub background: iced::Color,
-    pub border: Option<iced::Color>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct RadioButtons {
-    pub dot: iced::Color,
-    pub text: iced::Color,
-    pub border: iced::Color,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Sliders {
-    pub background: iced::Color,
-    pub border: iced::Color,
-    pub rail_border: Option<iced::Color>,
-    pub rail_backgrounds: (iced::Color, iced::Color),
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct ProgressBars {
-    pub background: iced::Color,
-    pub bar: iced::Color,
-    pub border: Option<iced::Color>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct PaneGrid {
-    pub background: iced::Color,
-    pub highlight_border: iced::Color,
-    pub highlight_background: iced::Color,
-    pub picked_split: iced::Color,
-    pub hovered_split: iced::Color,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Togglers {
-    pub on: Toggler,
-    pub off: Toggler,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct Toggler {
-    pub background: iced::Color,
-    pub background_border: iced::Color,
-    pub foreground: iced::Color,
-    pub foreground_border: iced::Color,
-}
-
-impl std::default::Default for Palette {
-    fn default() -> Self {
-        Self {
-            general: General {
-                background: color::LIGHT_BLACK,
-                foreground: color::BLACK,
-                scrollable: color::GREY_7,
-                accent: color::GREEN,
-            },
-            text: Text {
-                primary: color::WHITE,
-                secondary: color::GREY_2,
-                warning: color::ORANGE,
-                success: color::GREEN,
-                error: color::RED,
-                accent: color::BLUE,
-            },
-            buttons: Buttons {
-                primary: Button {
-                    active: ButtonPalette {
-                        background: color::GREEN,
-                        text: color::LIGHT_BLACK,
-                        border: color::GREEN.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::GREEN,
-                        text: color::LIGHT_BLACK,
-                        border: color::GREEN.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::GREEN,
-                        text: color::LIGHT_BLACK,
-                        border: color::GREEN.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREY_2,
-                        border: color::GREY_7.into(),
-                    }),
-                },
-                secondary: Button {
-                    active: ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREY_2,
-                        border: color::GREY_7.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREEN,
-                        border: color::GREEN.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREEN,
-                        border: color::GREEN.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREY_2,
-                        border: color::GREY_7.into(),
-                    }),
-                },
-                destructive: Button {
-                    active: ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::RED,
-                        border: color::RED.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::RED,
-                        text: color::LIGHT_BLACK,
-                        border: color::RED.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::RED,
-                        text: color::LIGHT_BLACK,
-                        border: color::RED.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::RED,
-                        border: color::RED.into(),
-                    }),
-                },
-                transparent: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    }),
-                },
-                transparent_border: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::GREEN.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::GREEN.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-                container: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: None,
-                    }),
-                },
-                container_border: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::GREEN.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::GREEN.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-                menu: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::WHITE,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::LIGHT_BLACK,
-                        text: color::WHITE,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::LIGHT_BLACK,
-                        text: color::WHITE,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::WHITE,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-                tab: Button {
-                    active: ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREY_2,
-                        border: color::GREY_7.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREEN,
-                        border: color::GREEN.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::LIGHT_BLACK,
-                        text: color::GREEN,
-                        border: color::GREEN.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::GREY_6,
-                        text: color::GREY_2,
-                        border: color::GREY_7.into(),
-                    }),
-                },
-                link: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREEN,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREEN,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::GREY_2,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-            },
-            cards: Cards {
-                simple: ContainerPalette {
-                    background: color::GREY_6,
-                    text: None,
-                    border: Some(color::TRANSPARENT),
-                },
-                modal: ContainerPalette {
-                    background: color::LIGHT_BLACK,
-                    text: None,
-                    border: color::TRANSPARENT.into(),
-                },
-                border: ContainerPalette {
-                    background: color::TRANSPARENT,
-                    text: None,
-                    border: color::GREY_7.into(),
-                },
-                invalid: ContainerPalette {
-                    background: color::LIGHT_BLACK,
-                    text: color::RED.into(),
-                    border: color::RED.into(),
-                },
-                warning: ContainerPalette {
-                    background: color::LIGHT_BLACK,
-                    text: color::ORANGE.into(),
-                    border: color::ORANGE.into(),
-                },
-                error: ContainerPalette {
-                    background: color::LIGHT_BLACK,
-                    text: color::RED.into(),
-                    border: color::RED.into(),
-                },
-            },
-            banners: Banners {
-                network: ContainerPalette {
-                    background: color::BLUE,
-                    text: color::LIGHT_BLACK.into(),
-                    border: None,
-                },
-                warning: ContainerPalette {
-                    background: color::ORANGE,
-                    text: color::LIGHT_BLACK.into(),
-                    border: None,
-                },
-            },
-            badges: Badges {
-                simple: ContainerPalette {
-                    background: color::GREY_4,
-                    text: None,
-                    border: color::TRANSPARENT.into(),
-                },
-                bitcoin: ContainerPalette {
-                    background: color::ORANGE,
-                    text: color::WHITE.into(),
-                    border: color::TRANSPARENT.into(),
-                },
-            },
-            pills: Pills {
-                primary: ContainerPalette {
-                    background: color::GREEN,
-                    text: color::LIGHT_BLACK.into(),
-                    border: color::TRANSPARENT.into(),
-                },
-                simple: ContainerPalette {
-                    background: color::TRANSPARENT,
-                    text: color::GREY_3.into(),
-                    border: color::GREY_3.into(),
-                },
-                warning: ContainerPalette {
-                    background: color::TRANSPARENT,
-                    text: color::RED.into(),
-                    border: color::RED.into(),
-                },
-                success: ContainerPalette {
-                    background: color::GREEN,
-                    text: color::LIGHT_BLACK.into(),
-                    border: color::GREEN.into(),
-                },
-            },
-            notifications: Notifications {
-                pending: ContainerPalette {
-                    background: color::GREEN,
-                    text: color::LIGHT_BLACK.into(),
-                    border: Some(color::GREEN),
-                },
-                error: ContainerPalette {
-                    background: color::ORANGE,
-                    text: color::LIGHT_BLACK.into(),
-                    border: Some(color::ORANGE),
-                },
-            },
-            text_inputs: TextInputs {
-                primary: TextInput {
-                    active: TextInputPalette {
-                        background: color::TRANSPARENT,
-                        icon: color::TRANSPARENT,
-                        placeholder: color::GREY_7,
-                        value: color::GREY_2,
-                        selection: color::GREEN,
-                        border: Some(color::GREY_7),
-                    },
-                    disabled: TextInputPalette {
-                        background: color::TRANSPARENT,
-                        icon: color::TRANSPARENT,
-                        placeholder: color::GREY_7,
-                        value: color::GREY_2,
-                        selection: color::GREEN,
-                        border: Some(color::GREY_7),
-                    },
-                },
-                invalid: TextInput {
-                    active: TextInputPalette {
-                        background: color::TRANSPARENT,
-                        icon: color::TRANSPARENT,
-                        placeholder: color::GREY_7,
-                        value: color::GREY_2,
-                        selection: color::GREEN,
-                        border: Some(color::RED),
-                    },
-                    disabled: TextInputPalette {
-                        background: color::TRANSPARENT,
-                        icon: color::TRANSPARENT,
-                        placeholder: color::GREY_7,
-                        value: color::TRANSPARENT,
-                        selection: color::GREEN,
-                        border: Some(color::RED),
-                    },
-                },
-            },
-            checkboxes: Checkboxes {
-                icon: color::GREEN,
-                text: color::GREY_2,
-                background: color::GREY_4,
-                border: Some(color::GREY_4),
-            },
-            radio_buttons: RadioButtons {
-                dot: color::GREEN,
-                text: color::GREY_2,
-                border: color::GREY_7,
-            },
-            sliders: Sliders {
-                background: color::GREEN,
-                border: color::GREEN,
-                rail_border: None,
-                rail_backgrounds: (color::GREEN, color::GREY_7),
-            },
-            progress_bars: ProgressBars {
-                bar: color::GREEN,
-                border: color::TRANSPARENT.into(),
-                background: color::GREY_6,
-            },
-            rule: color::GREY_1,
-            pane_grid: PaneGrid {
-                background: color::BLACK,
-                highlight_border: color::GREEN,
-                highlight_background: color::TRANSPARENT_GREEN,
-                picked_split: color::GREEN,
-                hovered_split: color::GREEN,
-            },
-            togglers: Togglers {
-                on: Toggler {
-                    background: color::GREEN,
-                    background_border: color::GREEN,
-                    foreground: color::WHITE,
-                    foreground_border: color::WHITE,
-                },
-                off: Toggler {
-                    background: color::GREY_2,
-                    background_border: color::GREY_2,
-                    foreground: color::WHITE,
-                    foreground_border: color::WHITE,
-                },
-            },
-        }
-    }
-}
-
-impl Palette {
-    /// Business theme: Light mode with cyan-blue accent
-    pub fn business() -> Self {
-        Self {
-            general: General {
-                background: color::LIGHT_BG,
-                foreground: color::LIGHT_BG_SECONDARY,
-                scrollable: color::LIGHT_BORDER,
-                accent: color::BUSINESS_BLUE,
-            },
-            text: Text {
-                primary: color::DARK_TEXT_PRIMARY,
-                secondary: color::DARK_TEXT_SECONDARY,
-                warning: color::ORANGE,
-                success: color::DARK_GREEN,
-                error: color::RED,
-                accent: color::BUSINESS_BLUE_DARK,
-            },
-            buttons: Buttons {
-                primary: Button {
-                    active: ButtonPalette {
-                        background: color::BUSINESS_BLUE,
-                        text: color::WHITE,
-                        border: color::BUSINESS_BLUE.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::BUSINESS_BLUE_DARK,
-                        text: color::WHITE,
-                        border: color::BUSINESS_BLUE_DARK.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::BUSINESS_BLUE_DARK,
-                        text: color::WHITE,
-                        border: color::BUSINESS_BLUE_DARK.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::LIGHT_BORDER_STRONG,
-                        text: color::WHITE,
-                        border: color::LIGHT_BORDER_STRONG.into(),
-                    }),
-                },
-                secondary: Button {
-                    active: ButtonPalette {
-                        background: color::LIGHT_BLUE_TINT,
-                        text: color::DARK_TEXT_SECONDARY,
-                        border: color::LIGHT_BORDER.into(), // Neutral border, blue appears on hover
-                    },
-                    hovered: ButtonPalette {
-                        background: color::LIGHT_BLUE_TINT, // Keep same background as active (like liana)
-                        text: color::BUSINESS_BLUE,
-                        border: color::BUSINESS_BLUE.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::LIGHT_BLUE_TINT, // Keep same background
-                        text: color::BUSINESS_BLUE,
-                        border: color::BUSINESS_BLUE.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::LIGHT_BG_TERTIARY,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: color::LIGHT_BORDER.into(),
-                    }),
-                },
-                destructive: Button {
-                    active: ButtonPalette {
-                        background: color::LIGHT_BG_SECONDARY,
-                        text: color::RED,
-                        border: color::RED.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::RED,
-                        text: color::WHITE,
-                        border: color::RED.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::RED,
-                        text: color::WHITE,
-                        border: color::RED.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::LIGHT_BG_TERTIARY,
-                        text: color::RED,
-                        border: color::RED.into(),
-                    }),
-                },
-                transparent: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY, // Use primary (black) for better visibility
-                        border: None,
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE, // Blue on hover
-                        border: None,
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: None,
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: None,
-                    }),
-                },
-                transparent_border: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY, // Use primary (black) for better visibility
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: color::BUSINESS_BLUE.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: color::BUSINESS_BLUE.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-                container: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_SECONDARY,
-                        border: None,
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY,
-                        border: None,
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY,
-                        border: None,
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY,
-                        border: None,
-                    }),
-                },
-                container_border: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY, // Use primary (black) for better visibility
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: color::BUSINESS_BLUE.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: color::BUSINESS_BLUE.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-                menu: Button {
-                    active: ButtonPalette {
-                        background: color::LIGHT_BG_SECONDARY,
-                        text: color::DARK_TEXT_PRIMARY,
-                        border: color::LIGHT_BG_SECONDARY.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::LIGHT_BG_TERTIARY,
-                        text: color::DARK_TEXT_PRIMARY,
-                        border: color::LIGHT_BG_TERTIARY.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::BUSINESS_BLUE,
-                        text: color::WHITE,
-                        border: color::BUSINESS_BLUE.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::LIGHT_BG_SECONDARY,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: color::LIGHT_BG_SECONDARY.into(),
-                    }),
-                },
-                tab: Button {
-                    active: ButtonPalette {
-                        background: color::LIGHT_BG_SECONDARY,
-                        text: color::DARK_TEXT_PRIMARY, // Use primary (black) for better visibility
-                        border: color::LIGHT_BORDER.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::LIGHT_BG_SECONDARY,
-                        text: color::BUSINESS_BLUE_DARK,
-                        border: color::BUSINESS_BLUE_DARK.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::LIGHT_BG,
-                        text: color::BUSINESS_BLUE_DARK,
-                        border: color::BUSINESS_BLUE_DARK.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::LIGHT_BG_TERTIARY,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: color::LIGHT_BORDER.into(),
-                    }),
-                },
-                link: Button {
-                    active: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_PRIMARY, // Use primary (black) for better visibility
-                        border: color::TRANSPARENT.into(),
-                    },
-                    hovered: ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: color::TRANSPARENT.into(),
-                    },
-                    pressed: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::BUSINESS_BLUE,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                    disabled: Some(ButtonPalette {
-                        background: color::TRANSPARENT,
-                        text: color::DARK_TEXT_TERTIARY,
-                        border: color::TRANSPARENT.into(),
-                    }),
-                },
-            },
-            cards: Cards {
-                simple: ContainerPalette {
-                    background: color::LIGHT_BG_SECONDARY,
-                    text: None,
-                    border: Some(color::TRANSPARENT),
-                },
-                modal: ContainerPalette {
-                    background: color::LIGHT_BG,
-                    text: None,
-                    border: color::LIGHT_BORDER.into(),
-                },
-                border: ContainerPalette {
-                    background: color::TRANSPARENT,
-                    text: None,
-                    border: color::LIGHT_BORDER.into(),
-                },
-                invalid: ContainerPalette {
-                    background: color::LIGHT_BG,
-                    text: color::RED.into(),
-                    border: color::RED.into(),
-                },
-                warning: ContainerPalette {
-                    background: color::LIGHT_BG,
-                    text: color::ORANGE.into(),
-                    border: color::ORANGE.into(),
-                },
-                error: ContainerPalette {
-                    background: color::LIGHT_BG,
-                    text: color::RED.into(),
-                    border: color::RED.into(),
-                },
-            },
-            banners: Banners {
-                network: ContainerPalette {
-                    background: color::BUSINESS_BLUE,
-                    text: color::WHITE.into(),
-                    border: None,
-                },
-                warning: ContainerPalette {
-                    background: color::ORANGE,
-                    text: color::WHITE.into(),
-                    border: None,
-                },
-            },
-            badges: Badges {
-                simple: ContainerPalette {
-                    background: color::LIGHT_BG_TERTIARY,
-                    text: None,
-                    border: color::TRANSPARENT.into(),
-                },
-                bitcoin: ContainerPalette {
-                    background: color::ORANGE,
-                    text: color::WHITE.into(),
-                    border: color::TRANSPARENT.into(),
-                },
-            },
-            pills: Pills {
-                primary: ContainerPalette {
-                    background: color::BUSINESS_BLUE,
-                    text: color::WHITE.into(),
-                    border: color::TRANSPARENT.into(),
-                },
-                simple: ContainerPalette {
-                    background: color::LIGHT_BLUE_TINT,
-                    text: color::BUSINESS_BLUE_DARK.into(),
-                    border: color::SOFT_BLUE.into(),
-                },
-                warning: ContainerPalette {
-                    background: color::AMBER,
-                    text: color::WHITE.into(),
-                    border: color::AMBER.into(),
-                },
-                success: ContainerPalette {
-                    background: color::DARK_GREEN,
-                    text: color::WHITE.into(), // White text on dark green for contrast
-                    border: color::DARK_GREEN.into(),
-                },
-            },
-            notifications: Notifications {
-                pending: ContainerPalette {
-                    background: color::BUSINESS_BLUE,
-                    text: color::WHITE.into(),
-                    border: Some(color::BUSINESS_BLUE),
-                },
-                error: ContainerPalette {
-                    background: color::ORANGE,
-                    text: color::WHITE.into(),
-                    border: Some(color::ORANGE),
-                },
-            },
-            text_inputs: TextInputs {
-                primary: TextInput {
-                    active: TextInputPalette {
-                        background: color::LIGHT_BG,
-                        icon: color::DARK_TEXT_TERTIARY,
-                        placeholder: color::DARK_TEXT_TERTIARY,
-                        value: color::DARK_TEXT_PRIMARY,
-                        selection: color::BUSINESS_BLUE,
-                        border: Some(color::LIGHT_BORDER),
-                    },
-                    disabled: TextInputPalette {
-                        background: color::LIGHT_BG_TERTIARY,
-                        icon: color::DARK_TEXT_TERTIARY,
-                        placeholder: color::DARK_TEXT_TERTIARY,
-                        value: color::DARK_TEXT_SECONDARY,
-                        selection: color::BUSINESS_BLUE,
-                        border: Some(color::LIGHT_BORDER),
-                    },
-                },
-                invalid: TextInput {
-                    active: TextInputPalette {
-                        background: color::LIGHT_BG,
-                        icon: color::DARK_TEXT_TERTIARY,
-                        placeholder: color::DARK_TEXT_TERTIARY,
-                        value: color::DARK_TEXT_PRIMARY,
-                        selection: color::BUSINESS_BLUE,
-                        border: Some(color::RED),
-                    },
-                    disabled: TextInputPalette {
-                        background: color::LIGHT_BG_TERTIARY,
-                        icon: color::DARK_TEXT_TERTIARY,
-                        placeholder: color::DARK_TEXT_TERTIARY,
-                        value: color::TRANSPARENT,
-                        selection: color::BUSINESS_BLUE,
-                        border: Some(color::RED),
-                    },
-                },
-            },
-            checkboxes: Checkboxes {
-                icon: color::BUSINESS_BLUE,
-                text: color::DARK_TEXT_PRIMARY,
-                background: color::LIGHT_BG_SECONDARY,
-                border: Some(color::LIGHT_BORDER),
-            },
-            radio_buttons: RadioButtons {
-                dot: color::BUSINESS_BLUE,
-                text: color::DARK_TEXT_PRIMARY,
-                border: color::LIGHT_BORDER,
-            },
-            sliders: Sliders {
-                background: color::BUSINESS_BLUE,
-                border: color::BUSINESS_BLUE,
-                rail_border: None,
-                rail_backgrounds: (color::BUSINESS_BLUE, color::LIGHT_BORDER),
-            },
-            progress_bars: ProgressBars {
-                bar: color::BUSINESS_BLUE,
-                border: color::TRANSPARENT.into(),
-                background: color::LIGHT_BG_TERTIARY,
-            },
-            rule: color::LIGHT_BORDER,
-            pane_grid: PaneGrid {
-                background: color::LIGHT_BG_SECONDARY,
-                highlight_border: color::BUSINESS_BLUE,
-                highlight_background: color::TRANSPARENT_BUSINESS_BLUE,
-                picked_split: color::BUSINESS_BLUE,
-                hovered_split: color::BUSINESS_BLUE,
-            },
-            togglers: Togglers {
-                on: Toggler {
-                    background: color::BUSINESS_BLUE,
-                    background_border: color::BUSINESS_BLUE,
-                    foreground: color::WHITE,
-                    foreground_border: color::WHITE,
-                },
-                off: Toggler {
-                    background: color::LIGHT_BORDER,
-                    background_border: color::LIGHT_BORDER,
-                    foreground: color::WHITE,
-                    foreground_border: color::WHITE,
-                },
-            },
-        }
-    }
-}
